+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - **sumneko:** removed lua-dev in favor of cmp-nvim-lua and removed buggy ui-select from telescope
 - **sumneko:** lua-dev only runs on neovim config path
 - **surfer:** more sane keybinding
+- **whichkey:** moved command to lua and fix source init lua
 
 > Revert
 - **neogen:** removed neogen